Update server/index with basic project setup
